Resolve relative --project-root for ai-plugin too
thv ai-plugin takes the same --project-root flag as thv skill and
validates it through the same code — pkg/plugins re-exports
skills.ValidateProjectRoot — so it had the identical bug: a relative
value travelled to the API server and came back naming the wire field
rather than the flag the user typed.

Move absProjectRoot out of skill_helpers.go into its own file now that
both command surfaces share it, and apply it to install, info, list, and
uninstall.

Follows #6211.
