+++ cmd/thv/app/skill_helpers.go
@@ -7,7 +7,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"path/filepath"
 
 	"github.com/spf13/cobra"
 
@@ -66,24 +65,6 @@ func validateProjectRootForScope(scopeVar, projectRootVar *string) func(*cobra.C
 		}
 		return nil
 	}
-}
-
-// absProjectRoot makes a user-supplied --project-root absolute, leaving an
-// empty value empty — for most commands that means "not project-scoped", and
-// substituting the working directory there would silently change the scope.
-//
-// The API server requires an absolute project root, so a relative value would
-// otherwise travel all the way to the server and come back as a validation
-// error naming the wire field rather than the flag the user typed.
-func absProjectRoot(explicit string) (string, error) {
-	if explicit == "" {
-		return "", nil
-	}
-	abs, err := filepath.Abs(explicit)
-	if err != nil {
-		return "", fmt.Errorf("resolving --project-root %q: %w", explicit, err)
-	}
-	return abs, nil
 }
 
 // resolveProjectRoot returns explicit if set, otherwise auto-detects the
